backend/pyproject.toml
# ==============================================================================
# PROJECT METADATA & CORE DEPENDENCIES
# ==============================================================================
[project]
name = "wedding-management"
version = "0.1.0"
description = "Sistema de gestão de casamentos - Backend Django REST API"
requires-python = ">=3.12"
dependencies = [
    # Core Django
    "django>=5.2.9,<6.0",
    "djangorestframework>=3.16.1,<3.17",
    "djangorestframework-simplejwt>=5.4.0,<5.5",
    "django-cors-headers>=4.6.0,<4.7",
    "drf-spectacular>=0.27.1,<0.28",
    # Database
    "psycopg[c]>=3.2.0,<4.0",
    # Utils
    "uuid6>=2025.0.1,<2026",
    "python-dotenv>=1.0.1,<2.0",
    # Production Server
    "gunicorn>=23.0.0,<24.0",
    "pydantic>=2.12.5",
]

# ==============================================================================
# DEVELOPMENT DEPENDENCY GROUPS (UV Native)
# ==============================================================================
[dependency-groups]
dev = [
    # Testing (Pytest Stack)
    "pytest>=8.3.4,<9.0",
    "pytest-django>=4.9.0,<5.0",
    "pytest-env>=1.2.0",
    "pytest-factoryboy>=2.8.1",
    "pytest-cov>=6.0.0,<7.0",
    "pytest-xdist>=3.6.1,<4.0",
    "pytest-mock>=3.14.0,<4.0",
    "factory-boy>=3.3.1,<3.4",
    "faker>=34.0.0,<35.0",
    # Development Tools
    "django-extensions>=4.1,<4.2",
    "django-zeal>=2.0.4,<3.0",
    "ipython>=9.8.0,<10.0",
    # Code Quality & Linting
    "ruff>=0.7.4,<0.8",
    "pre-commit>=3.7.1,<3.8",
    "detect-secrets>=1.5.0",
]

# ==============================================================================
# CODE QUALITY: RUFF CONFIGURATION
# ==============================================================================
[tool.ruff]
line-length = 88
target-version = "py312"
preview = true
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "venv",
    ".venv",
]

[tool.ruff.lint]
# Adicionada a regra 'B' explicitamente para pegar o erro B904
select = ["E", "F", "W", "I", "B", "RUF", "UP", "DJ"]
# I001 é o organize imports, não ignore se você quer que ele organize ao salvar!
ignore = ["E203", "RUF012"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
# Forçar line-ending para evitar problemas de buffer no Windows/Linux
line-ending = "lf"


# ==============================================================================
# TESTING: PYTEST CONFIGURATION
# ==============================================================================
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"

# Padrões de descoberta de arquivos
python_files = ["test_*.py", "*_test.py", "tests_*.py", "*_tests.py", "test.py", "tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Variáveis de ambiente isoladas para teste (Requer pytest-env)
env = [
    "DEBUG=False",
    "SECRET_KEY=test-key-safe-for-production",
    "EMAIL_BACKEND=django.core.mail.backends.locmem.EmailBackend",
    "DEFAULT_FROM_EMAIL=test@wedding.com",
    "DATABASE_URL=postgres://user:pass@localhost:5432/test_wedding_db"
]

# Marcadores para organização de execução
markers = [
    "slow: marca testes lentos (deselect com '-m \"not slow\"')",
    "integration: testes de integração",
    "unit: testes unitários",
    "functional: testes funcionais",
]

# Opções de execução padrão
addopts = [
    "--strict-markers",
    "--reuse-db",
    "-ra",
    "--tb=short",
    "--ds=config.settings",
]

norecursedirs = [
    "migrations", ".git", ".tox", "venv", ".venv", "__pycache__", "media", "static", "node_modules"
]

log_cli = false
log_level = "ERROR"

filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ==============================================================================
# COVERAGE REPORTING
# ==============================================================================
[tool.coverage.run]
relative_files = true
branch = true
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "manage.py",
    "*/config/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __str__",
    "def __repr__",
    "if settings.DEBUG",
    "raise NotImplementedError",
]
